4_AI_test/My Email Noti (news)/requirements.txt
fastapi
uvicorn
sqlalchemy
requests
beautifulsoup4
schedule
jinja2
python-multipart
python-dotenv
openai
